requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
audioread==3.1.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
decorator==5.2.1
Django==5.0
django-cors-headers==4.9.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.0
filelock==3.19.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.50.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.45.1
MarkupSafe==2.1.5
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.3
numba==0.62.1
numpy==1.26.4
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
packaging==25.0
platformdirs==4.5.0
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.0.1
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
# scipy==1.12.0
scipy==1.14.1

sniffio==1.3.1
soundfile==0.12.1
soxr==1.0.0
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.1+cpu
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
